Customer class
Completed the constructor method for the Customer class by adding two variables of customerID and password.
Created an instance of a customer which takes an array of accounts, customerID and a password.
Added two methods isPasswordCorrect, which checks is a user has entered the correct password and getCustomerID which returns current customers' ID.